+++ diff/definition_tkdiff.go
@@ -1,0 +1,29 @@
+package diff
+
+func defineTkDiff() *ToolDefinition {
+	leftArgs := func(temp, target string) []string {
+		return []string{target, temp}
+	}
+
+	rightArgs := func(temp, target string) []string {
+		return []string{temp, target}
+	}
+
+	return &ToolDefinition{
+		Kind:             TkDiff,
+		Url:              "https://sourceforge.net/projects/tkdiff/",
+		Cost:             Free,
+		AutoRefresh:      false,
+		IsMdi:            false,
+		SupportsText:     true,
+		RequiresTarget:   true,
+		BinaryExtensions: []string{},
+		Osx: OsSettings{
+			TargetLeftArguments:  leftArgs,
+			TargetRightArguments: rightArgs,
+			ExePaths: []string{
+				"/Applications/TkDiff.app/Contents/MacOS/tkdiff",
+			},
+		},
+	}
+}
