+++ diff/definition_tortoisemerge.go
@@ -1,0 +1,29 @@
+package diff
+
+func defineTortoiseMerge() *ToolDefinition {
+	leftArgs := func(temp, target string) []string {
+		return []string{target, temp}
+	}
+
+	rightArgs := func(temp, target string) []string {
+		return []string{temp, target}
+	}
+
+	return &ToolDefinition{
+		Kind:             TortoiseMerge,
+		Url:              "https://tortoisesvn.net/TortoiseMerge.html",
+		Cost:             Free,
+		AutoRefresh:      false,
+		IsMdi:            false,
+		SupportsText:     true,
+		RequiresTarget:   true,
+		BinaryExtensions: []string{},
+		Windows: OsSettings{
+			TargetLeftArguments:  leftArgs,
+			TargetRightArguments: rightArgs,
+			ExePaths: []string{
+				"%ProgramFiles%\\TortoiseSVN\\bin\\TortoiseMerge.exe",
+			},
+		},
+	}
+}
